Only removing psf archive if it exists
Former-commit-id: dca624456686cc78932180d2a5b5b8cd3bca99cd

+++ SHE_GST_GalaxyImageGeneration/python/SHE_GST_GalaxyImageGeneration/generate_images.py
@@ -391,7 +391,8 @@ def generate_image_group(image_group_phl, options):
 
     # Remove the now-unneeded PSF archive file
     psf_archive_hdulist.close()
-    os.remove(os.path.join(workdir, psf_archive_filename))
+    if os.path.exists(os.path.join(workdir, psf_archive_filename)):
+        os.remove(os.path.join(workdir, psf_archive_filename))
 
     return
